+++ src/main/java/tornadofx/Component.kt
@@ -47,7 +47,6 @@ import kotlin.properties.ReadOnlyProperty
 import kotlin.reflect.*
 
 interface Injectable
-interface InstanceScoped
 
 interface Configurable {
     val config: Properties