+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - CSS warning should not be issued in OSGi environment, since bundle activator installs CSS URL Handler
 - All shape builders accepts `Number` instead of `Double` so you can write `circle(10, 10, 5)` instead of `circle(10.0, 10.0, 5.0)`
 - ComboBox.validator moved to ComboBoxBase.validator to support ColorPicker and DatePicker as well
+- Removed InstanceScoped and removed it from Wizard. It was not needed.
 
 ## [1.7.4] - 2017-04-28